+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ##  [Unreleased]
 
+_tbd_
+
+
+##  [0.2] – 2018-02-02
+
 
 ### Added
 
@@ -46,4 +51,5 @@ Initial release
 -   Get metrics
 
 
-[Unreleased]: https://github.com/bheisig/checkmkwebapi/compare/0.1...HEAD
+[Unreleased]: https://github.com/bheisig/checkmkwebapi/compare/0.2...HEAD
+[0.2]: https://github.com/bheisig/checkmkwebapi/compare/0.1...0.2